% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sharpe.R
\name{calc_sharpe}
\alias{calc_sharpe}
\title{Sharpe ratio from output of simulation}
\usage{
calc_sharpe(backtest_results, days_pa = 365)
}
\arguments{
\item{backtest_results}{tibble of backtest results from simulation}

\item{days_pa}{number of days in calendar year.
A lenth-one numeric vector.}
}
\value{
Sharpe from backtest simulation.
A lenth-one numeric vector.
}
\description{
Sharpe ratio from output of simulation
}
\author{
John Gavin
}
