% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_symbols_stables_llama.R
\name{get_symbols_stables_llama}
\alias{get_symbols_stables_llama}
\title{defi llama stablecoins}
\usage{
get_symbols_stables_llama(
  url = "https://stablecoins.llama.fi/stablecoins?includePrices=true"
)
}
\arguments{
\item{url}{url to download \url{llama.fi} (default) stable coin symbols.
A length-one charector vector.
If null (default)}
}
\value{
character vector
}
\description{
defi llama stablecoins
}
\examples{
get_symbols_stables_llama()
}
\author{
John Gavin
}
