% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_equity_curve.R
\name{ggplot_equity_curve}
\alias{ggplot_equity_curve}
\title{Equity curve from output of backtest}
\usage{
ggplot_equity_curve(backtest_results, commission_pct, trade_on = "close")
}
\arguments{
\item{backtest_results}{Backtest results. A tibble}

\item{commission_pct}{Percent commission per trade.}

\item{trade_on}{Column name in \code{backtest_results}.
Real length-one vector.}
}
\value{
ggplot object
}
\description{
Equity curve is output
as a ggplot object to be plotted.
}
\author{
John Gavin
}
