% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robotwealth.com-package.R
\docType{package}
\name{robotwealth.com-package}
\alias{robotwealth.com}
\alias{robotwealth.com-package}
\title{robotwealth.com: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JohnGavin/robotwealth.com}
  \item Report bugs at \url{https://github.com/JohnGavin/robotwealth.com/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
