% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stablecoins}
\alias{stablecoins}
\title{Stablecoins from llama}
\format{
\subsection{\code{stablecoins}}{

A character vector of stablecoims names.
}
}
\source{
\url{https://stablecoins.llama.fi/stablecoins?includePrices=true}
}
\usage{
stablecoins
}
\description{
Stablecoins from llama
}
\keyword{datasets}
