% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_config_set_project.R
\name{tar_config_set_project}
\alias{tar_config_set_project}
\title{Adds new targets project to .yaml file}
\usage{
tar_config_set_project(
  pr_title = "conv_opt_cvxr",
  pr_yaml = paste0(pr_title, ".yaml"),
  pr_script = paste0(pr_title, ".R"),
  pr_store = paste0(pr_title, "_store")
)
}
\arguments{
\item{pr_title}{project title.}

\item{pr_yaml}{project yaml.}

\item{pr_script}{tar_script filename.}

\item{pr_store}{targets db to store objects.}
}
\value{
nothing returned.
}
\description{
All parameters are length-one character vectors.
Sets two \code{targets} environment variables:
\code{TAR_PROJECT} and \code{TAR_CONFIG}.
Sets target configuration via \code{tar_config_set}.
Parameters are stored in top-level project folder.
\code{./<pr_yaml>.yaml}.
}
\author{
John Gavin
}
